Link processes that are doing transactions
In the case of pool handling and supervisors a connection can be placed in a
invalid state if a process is using a checked out connection and then is
brutally killed by, for example, its supervisor. To avoid this a link between
the calling process and the connection is being kept while performing the
transaction so that the connection process is killed in case the calling process
dies.