Monitor calling process during transaction
Monitoring the caller allows the connection to kill itself if the
calling process is killed (#96)

When using Poolboy (or similar) for connection pooling and a when
process who has checked out a connection is killed, the connection
goes back into the pool, even if it is in an ongoing transaction.
Monitoring the calling process when a transaction is started, the
connection can kill itself if the caller is killed and thus prevent
connections in a bad state from being put back into the pool.