Handle error packet as the initial packet from the server
Some versions of MariaDB (e.g. 10.1.21) may send an error packet as
the first packet, e.g. in case of 'Too many connections'. This seems
to be a bug in some versions as it is not permitted according to the
protocol documentation. Nevertheless, nothing prevents us from
handling an error packet as the first message from the server.

Fixes #88