Possibility to apply a filtermap fun on results (#104)
The query/execute functions are provided
with an additional parameter to allow for an (optional) filtermap
function to be passed in, for filtering and/or mapping the rows in the results in a similar fashion as lists:filtermap/2.

This fun is applied in the process of the connection, so that copying to the caller process can be avoided for large result sets.