Disable the deadlock test for MySQL 5.7.x
This test has been seen failing for MySQL 5.7.25. When a deadlock
occurs, one of the transactions gets implictly committed by the
server. The expected behavior is a deadlock error, which is
handled by the client by restarting the transaction.

The test can be enabled for MySQL 5.7.x when it has been
identified whether this is a known bug for certain MySQL versions
and how to best handle that.

This commit also disables testing on Travis CI for Erlang R16B,
which doesn't seem to be available anymore.