Check logged errors in tests and hide them from the output
This was implemented before for error_logger, but the errors
were not hidden for OTP >= 21.0, where logger is used. Also,
some tests didn't use this mechanism so the logged errors
were printed also when running under older OTP releases.

With this change, the output from a successful test run is
tidy, so it is easy to spot if anything unexpected happens.