Override mode, packet and active TCP options
A user could provide TCP options that would make a connection unusable,
namely `list` or `{mode, list}`, `packet` with a value other than `0` or
`raw`, and `active` with a value other than `false`.

The changes in this commit override the aforementioned options with the
correct values, and makes `true` the default for the `nodelay` option
but allows a user to set it to something else if wanted.