Default SNI to hostname for SSL connections
The ssl option `server_name_indication` did not have a default value,
which, in conjunction with the `verify` option defaulting to `verify_peer`,
meant that the IP address will be used in order to check the identities
in the certificate.
This change sets the default value for `server_name_indication` to the
value of the `host` option given in `mysql:start_link/1` if it is a hostname
string (ie, not a IP address tuple or string or an atom). Otherwise, no
default value will be set.