allow prepare query to have a timeout
It's been a while but this patch was sitting in
one of my private gist. I believe the issue was
during one of the usage of mysql wherein the underlying
assumption was that timeout was applied but eventaully
there were a lot of prepare queries which ended up
in locking the table. Hence the application started
experiencing table lock errors when there were too many
prepare queries and MySQL was unable to cope up with it.
This patch attempts to fix such scenarios and fail with
timeout being applied.