Add 'decode_decimal' option to control how DECIMALs are returned (#194)
The new option `{decode_decimal, auto | number | float | binary}` controls how a
DECIMAL value is translated to an Erlang term. The default 'auto' is the legacy behaviour which returns an integer or a float when no precision loss can happen for the column and a binary otherwise. The options 'number' and 'float' may return a float even if it can result in precision loss, while 'binary' always returns the textual representation of the number.