Hide sensitive information in state and use `gen_server:format_status/1,2` depending on OTP release (#212)
We use the `gen_server:format_status/2` callback in order to hide
sensitive data from logs.

In OTP/25, another callback `format_status/1` was introduced which
should replace `format_status/2`. In OTP/27, `format_status/2` was
deprecated, and raises a compiler warning.

This PR adds the `format_status/1` callback, and a conditional that adds
the `format_status/2` callback for OTP<25, where `format_status/1` is
absent.

I also changed the behavior such that it _always_ hides the sensitive
data, ie not only in case of crashes, because BSTS ;) If a user _needs_
the unobfuscated state, he can always use `sys:get_state`.