Fix `public_key:encrypt_public` deprecation warning (#214)
`mysql_protocol` uses `public_key:encrypt_public`, which was deprecated
in OTP 27.0 for securtity reasons. This function was undeprecated again
in OTP 27.1 because only `rsa_pkcs1_padding` (used in MySQL < 8.0.5) is
unsafe, while `rsa_pkcs1_oaep_padding` (used in MySQL >= 8.0.5) is still
safe.

This commit works around the deprecation warning in OTP 27.0 and at the
same time logs a warning when `rsa_pkcs1_padding` is used.