Fix test for `socket` backend (#219)
The socket backend that `mysql_tests/socket_backend_test` tests is only
supported in OTP 23+, and still very unstable in 23. Therefore, the
change in this PR skips the test on OTP =< 23.

Also changed the level of logs emitted from the test suite from `info`
to `notice`, because `info` is below the logging threshold in the
default configuration, ie the minimum level is `notice`.